requirements.txt
appdirs==1.4.4
asgiref==3.3.2
astroid==2.6.2
asttokens==2.0.5
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2022.12.7
cfgv==3.3.0
charset-normalizer==2.0.8
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
distlib==0.3.2
Django==3.2.18
django-doctor==2.5.0
dparse==0.5.2
filelock==3.0.12
flake8==3.9.2
identify==2.2.11
idna==3.3
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.8.1
platformdirs==2.2.0
poetry-semver==0.1.0
pre-commit==2.13.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.9.3
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
rich==10.6.0
six==1.16.0
skjold==0.3.2
sqlparse==0.4.4
toml==0.10.2
tomli==1.0.4
urllib3==1.26.7
virtualenv==20.7.0
wrapt==1.12.1
django-filter==23.2
qrcode==7.4.2
